Convert to poetry and add Starship support
