+++ cli/pyproject.toml
@@ -1,28 +1,23 @@
-[project]
+[tool.poetry]
 name = "trak"
 version = "0.0.1"
-description = "Keep a record of the time you dedicate to your projects."
-authors = [
-    {name = "Luca Fedrizzi", email = "fedrizzi.luca@hotmail.it"},
-]
-dependencies = [
-    "typer[all]>=0.9.0",
-]
-requires-python = ">=3.11"
+description = "Keep a record of the time you dedicate to your projects"
+authors = ["lcfd <9001053+fedriz@users.noreply.github.com>"]
+license = "AGPL-3.0"
 readme = "README.md"
-license = {text = "AGPL-3.0"}
+
+[tool.poetry.scripts]
+trak = "trak.main:app"
+
+[tool.poetry.dependencies]
+python = "^3.11"
+typer = {extras = ["all"], version = ">=0.9.0"}
+
+
+[tool.poetry.group.dev.dependencies]
+black = ">=23.9.1"
+ruff = "^0.0.292"
 
 [build-system]
-requires = ["pdm-backend"]
-build-backend = "pdm.backend"
-
-[tool.pdm.dev-dependencies]
-style = [
-    "black>=23.9.1",
-]
-test = [
-    "pytest>=7.4.2",
-]
-lint = [
-    "ruff>=0.0.292",
-]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
