--- cli/src/tests/test_database.py
@@ -1,1 +1,0 @@
-
