--- cli/src/tests/test_trak.py
@@ -1,10 +1,0 @@
-from typer.testing import CliRunner
-
-from trak import cli
-
-runner = CliRunner()
-
-
-def test_version():
-    result = runner.invoke(cli.app, ["--version"])
-    assert result.exit_code == 0
