+++ cli/trak/__main__.py
@@ -1,0 +1,5 @@
+from trak import __app_name__
+
+from trak.main import app
+
+app(prog_name=__app_name__)
