+++ cli/trak/config.py
@@ -1,6 +1,5 @@
 from pathlib import Path
 
-
 #
 # Paths
 #
