+++ cli/trak/utils/print_with_padding.py
@@ -1,0 +1,5 @@
+from rich.padding import Padding
+
+
+def print_with_padding(text: str, x: int = 2, y: int = 2):
+    return Padding(text, (y, x))
